requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-gis>=1.0
django-filter>=23.0
django-allauth>=65.0
django-cors-headers>=4.0
django-jazzmin>=3.0
requests>=2.31
Pillow>=10.0
